+++ js/api.js
@@ -464,7 +464,9 @@
 		if (action !== false) {
 			$e_initializeUISetup();
 			// Apply to currently running code if any
-			$_eseecode.execution.endLimit = $_eseecode.execution.startTime+value*1000;
+			if ($_eseecode.execution.startTime !== undefined) {
+				$_eseecode.execution.endLimit = $_eseecode.execution.startTime+value*1000;
+			}
 		}
 	}